Move `Format` trait to `utils` and add documentation.
